+++ webapp/frontend/src/app/features/channels/channels.component.html
@@ -2,7 +2,7 @@
   <heading title="Подкасты"></heading>
 
   <div class="channels__filter">
-    <channels-filter></channels-filter>
+    <channels-filter [channels]="mockChannelsList"></channels-filter>
   </div>
 
   <ng-container *ngFor="let card of mockCards">
