+++ webapp/frontend/src/app/features/channels/components/channels-filter/channels-filter.component.scss
@@ -19,36 +19,3 @@
     align-items: center;
   }
 }
-
-// .mat-form-field {
-//   font-family: $font-family !important;
-// }
-
-// .mat-primary .mat-option.mat-selected:not(.mat-option-disabled) {
-//   color: $black-color;
-//   font-weight: 500;
-// }
-
-// .mat-form-field-infix {
-//   min-width: 10rem !important;
-// }
-
-// .mat-form-field-appearance-fill .mat-form-field-flex {
-//   background-color: #fff;
-//   border: none;
-//   border: 1px solid $gray-color;
-//   border-radius: 0.5rem !important;
-//   padding: 0 0.75em !important;
-// }
-
-// .mat-form-field-underline {
-//   display: none;
-// }
-
-// .mat-form-field-appearance-fill .mat-select-arrow-wrapper {
-//   transform: translateY(0) !important;
-// }
-
-// .mat-form-field-wrapper {
-//   padding-bottom: 0;
-// }
