+++ webapp/frontend/src/app/features/channels/components/podcast-card/podcast-card.component.scss
@@ -56,7 +56,5 @@
 
   &__player {
     width: 100%;
-    // height: 1rem;
-    // background-color: #333;
   }
 }
