+++ webapp/frontend/src/app/features/channels/components/slider/slider.component.html
@@ -12,7 +12,6 @@
           <ng-template [ngIf]="channel.logo">
             <i class="logo" [ngStyle]="{ backgroundImage: 'url(' + channel.logo + ')' }"></i>
           </ng-template>
-
           <p>{{ channel.title }}</p>
         </span>
       </ng-template>
