+++ webapp/frontend/src/app/features/channels/model/slider-item.model.ts
@@ -1,0 +1,5 @@
+import { Channel } from './channel.model';
+
+export interface SliderItem extends Channel {
+  id?: string;
+}
