+++ webapp/frontend/src/app/shared/components/player/player.component.scss
@@ -2,8 +2,6 @@
 
 .player {
   width: 100%;
-  //height: 1rem;
-  //background-color: #333;
   display: flex;
   flex-direction: row;
   justify-content: stretch;
